Make simplifier and Core Lint join-point-aware
New command-line options, -ffloat-joins-only-to-top and -fcontext-substitution,
make the simplifier smarter when handling join points.

Also, JoinPointInfo is no longer a synonym for Bool. If something is a join
point, we need to know exactly how many arguments it has *as a join point* -
its call sites may be partial applications (though they must be consistently
so), in which case its "join point arity" is lower than the number of lambdas it
starts with.

Core Lint currently only checks join points when checking the output of the
simplifier or the Float Out pass.