Ignore joins entirely when context subst off
I'd been trying to allow fine control so that we could track join points but not
try hard to preserve them. In particular, I was allowing that we might float a
join point out of a value binding, so long as we decontify it. Problem is,
promoting a join point is only valid if it doesn't itself refer to other join
points. In retrospect, this shouldn't be a problem because there shouldn't be a
join boundary in the way anyway, but for the moment this is the simpler solution
anyway.

The upshot is that -ffloat-joins-only-to-top is now only relevant in the Float
Out pass; the simplifier just erases all join points if -fcontext-substitution
is off, then unconditionally deals carefully with any join points it finds.