Only mark join points as join points if sm_context_subst on
Since sm_context_subst's effect is to strip the join point information off of
everything, the rest of the simplifier assumes that any join point should be
handled as such. Thus to keep the simplifier to pre-CoreJoins behavior, we need
to zap the join point information of *new* binders as well.