Don't zap join points during iteration
We were saying a join point gets ruined if the number of lambdas on the outside
changes. I think this was because JoinPointInfo was still a boolean. Not only
was zapping the join point info unnecessary, it was harmful, because it causes
the simplifier to forget that it moved the outer context inside.