Track JoinPointInfo for vars being substituted
CoreJoins only sets the join point info at the binding site, so we can't rely on
the occurrences to figure out which function calls are to join points (crucial
because then we throw out the outer context!). We can look up the up-to-date
JoinPointInfo in the in-scope set, except that not all variables are in the
in-scope set: some are in the substitution instead! For most purposes, we don't
need up-to-date IdInfo for something we're substituting away anyway, but
JoinPointInfo is different, since it affects how the substitution proceeds (do
we throw out the outer context or not?). Hence there's now an seSubstJoins
field in SimplEnv (part of the static part) that remembers which variables in
the domain of seIdSubst are join points and with what arity. The new
joinVarArity function figures out what an incoming variable's JoinPointInfo is
by looking both in seSubstJoins *and* in the in-scope set.

(Arguably, this is all symptomatic of the abuse of IdInfo for highly
semantically relevant data. OTOH, it only has such strong relevance in the
simplifier.)

Relatedly, this patch also stops trimming the context when substituting
something that was pre-inlined, since its context hasn't been copied inside yet
(since it hasn't been simplified yet).