Use, fix simplExprFV
The floating joins are always associated with a very specific continuation, so
whenever the continuation changes, there must be a join float barrier. In
retrospect, this should have been obvious. simplExprFV provides a join float
barrier, and it is now called when entering an applied function or scrutinized
expression. Also, it has been fixed in the case where there are existing floats
for the outer continuation; these need to be zapped going in and reattached to
the environment coming out.