Improve eta-expansion of functions with recursive join points
Previously we adjusted the type of a join binder after pushing the new context
(with the args and coercions from eta-expansion) inside it, by rather
ham-handedly calling exprType on the new body. Now we calculate the new type
ahead of time so that occurrences of recursive join points have the right type.