Join points don't float out
This took me too long to realize, but yes, in the simplifier, join points never
float out (with the *possible* exception of a CAF with a top-level join point
for some reason). This simplifies some of the handling needed; now, the only
places that handle floating join points know exactly where they're coming from
(because they're calling simplNonRecE or prepareCaseCont, for example). There's
no need to block join points from floating out of terms because they never make
it that far to begin with, so no more simplExprFV.

In fact, it's a bit silly to keep the join points in the Floats structure at
all. Keeping them in the SimplEnv is very convenient the way the simplifier
handles bindings (anything taking a Bind returns only a SimplEnv), but there's
really no need to cram them together with the Floats because everyone either
calls wrapJoinFloats and zapJoinFloats or calls wrapFloats and zapFloats
expecting no join points.