Don't try to eta-expand recursive join points
We can't eta-expand a recursive join point, since afterward the recursive call
is necessarily out of the join point's scope (now that it's nestled in some
extra lambdas). This *could* be fixed by changing the join point's arity, but
that requires altering all the call sites (inside and outside), which is going
to be painful. I've added a WARN so we can see how often we miss an
eta-expansion due to this change.