Look for join points going bottom-up
We can find more join points if we don't assume that each one's call arity is
the same as the number of lambdas it starts with. This requires a bit more
intelligence of the traversal, since it has to check consistency among all the
call sites rather than just checking each one against the number of lambdas.

This patch doesn't actually let the number of lambdas differ from the call
arity, though, since there are subtleties that follow from that.