Improvements to CoreJoins
We no longer get tripped up by would-be join points that appear unsaturated in
the RULES of other join points.

  let join j1 x y = x + y
      join j2 x y = j1 y x
      {-# RULES "j2/0" j2 0 = j1 1 #-}
  in j2 0 4

Here the rule has an unsaturated occurrence of j1, *but* when the rule is
applied, the occurrence becomes saturated. Hence j1 and j2 can both be join
points.

Also, we don't make something a join point if it would violate the let/app
invariant. I intend to remove the invariant for join points, but for now let's
just not break things.