Float In: Don't float a join point into a call site
Float In would float a join point all the way down to the individual variable if
it only occurred once (common for recursive join points). This is the only way
for Float In to destroy a join point, since it only tries to float things where
there are occurrences, and there can't be occurrences anywhere a join point
isn't allowed to float.