Rescind the let/app invariant for join points
A nullary join point might have an unlifted type, but it is still essentially a
function - it will not be computed unless the join point is invoked (in this
case by a simple occurrence in tail position). Currently, join points created by
the simplifier always have a Void# parameter added, which, in the unlifted case,
is crucial to avoid falling under the let/app invariant. This patch renders this
superfluous (to be removed soon). Importantly, this puts nullary join points
that occur naturally on the same footing as those created by the simplifier.
(The desugarer can probably stop adding Void# arguments as well.)

Since code generation still expects the let/app invariant to hold, CorePrep now
adds a Void# parameter to any join point with no value parameters.

Places where the relaxation of the let/app invariant has an effect (undoubtedly
there are more):

  * CoreJoins - We can now make a join point from an unlifted let. Previously we
    couldn't, because then the simplifier might copy some context inside, making
    it no longer okay for speculation.

  * FloatIn - We now need to check whether a let-binding is a join point, since
    we may have no argument to check for one-shot-ness. Fortunately, the story
    is very simple for floating into a join point: don't do it if it's
    recursive, do it if it isn't (as then the value will be computed each time
    through the loop!).