Worker/Wrapper: A nullary join point is not a thunk
The W/W transformation on thunks turns

  x = e

into

  x = let x = e
      in case x of
           I# y -> let x = I# y in x,

which breaks if x is actually a nullary join point:

  join x = let join x = e
           in case x of -- oops! not a tail call
                I# y -> let join x = I# y in x

since the occurrence of x in the case is not a tail call. We could of course
pick a fresh non-join variable for the local:

  join x = let v = e
           in case v of
                I# y -> let join x = I# y in x

But now if e calls *another* join point, that call is no longer valid (since
join points cannot be called from value bindings).

The solution here is simply not to do the transformation in this case. In the
same way that CPR is unnecessary on join points, some outer case is very likely
to move inward and scrutinize e anyway.