Remove propagateBinders
This was a highly sketchy way to deal with join points that might get
decontified (because of the command-line flag that lets us float them out
uncarefully to measure how bad that is). Only Float Out used it, and it already
handles decontification by zapping floated things and updating call sites.