DmdAnal: Joins aren't thunks
CPR analysis was treating nullary join points as thunks and thus not giving them
the CPR property. Since join points don't get CPR'd, this is only a problem if
the CPR property would have propagated to something else as well, but it did
cause a ~4% increase in allocs in boyer2.