Simplify: Adjust type of join point before simplifying it
Previously, we were simplifying the RHS, then using the simplified version to
determine the new type. This is a problem if the join point is recursive,
since the binder goes into the environment with the wrong type. However, we can
determine the new type from the original type and the continuation we're pushing
into the RHS, so now we change the type before the binder gets put into the
environment.