Float In: Don't float a function into its own call
There's really no point in turning

  let f = ...
  in f 1 2 3

into

  (let f = ...
  in f) 1 2 3

because the simplifier will put it right back. Until join points, however, there
was no harm in doing so either. Nowadays, f can be made a join point in the
former case but not the latter. This is a problem if f is being floated inward
from farther away; if we don't stop floating at the right place, we won't
discover that it's a join point until after a pass of the simplifier (which may
already do something differently if f starts out as a join point).