Float In, Float Out: Find more join points
Tweaking the pipeline a bit in the hopes of finding more join points. See
test join002 for code that should be improved here. The idea is to run Float In
soon after the initial gentle simplification, since here it has a good chance of
floating bindings into the right place (say, inside the scrutinee of a case) so
they can be made into join points. Then, we stop Float Out from floating join
points to top level; I'd been having it do this because a top-level function
isn't much slower than a join point and maybe we can leave something inlinably
small behind, but we lose the ability to move context into the join point in the
simplifier. Added -fallow-joins-to-float-to-top so we can switch and measure.