Specialise: Specialise join points as join points
This actually requires cooperation from CoreLint and the simplifier, since the
generated rules generally don't match all the arguments to the join point,
meaning the rule's RHS is an unsaturated call to a join point, which is normally
verboten. However, since the RHS will only be substituted into contexts which
supply the extra arguments, in the end no invariants will be broken if we're
careful.

The simplifier deals with a lower-arity rule for a join point simply by
"eta-expanding" the rule before proceeding; thus it still only deals with
saturated calls to join points.