CoreArity: Throw out any "reflexive" coercion
For pushing eta-expansions into join points, it's often important that a
coercion cancel out its opposite so that the cast doesn't end up between a join
binding and an occurrence. There's no reason to keep a coercion from a type to
itself, so we check for this and toss the coercion (just as the simplifier
does).