CorePrep: Fix adding void parameters to join points
Rather than inferring which join points are getting void parameters, we now
remember in the environment, which is much more reliable. The issue that forced
this was that a join point might have arity 0 yet return a function with a value
argument; thus it does *not* need the Void# parameter, but you can't figure that
out from the occurrence site alone.