Overhaul Float Out for join points
Join points should be able to float outward, just not very far. An important use
for this appears in unfold/destroy fusion: when several filters fuse together,
we wind up with nested loops that don't actually do anything:

  let join go1 x y =
    let join go2 x y =
      let join go3 x y = ...
      in go3 x y
    in go2 x y
  in go1 x y

If we let Float Out do its thing, we get:

  let join go3 x y = ...
           go2 x y = go3 x y
           go1 x y = go2 x y
  in go1 x y

Which of course collapses down to a single loop once inlining takes place:

  let join go3 x y = ...
  in go3 x y

Unfortunately, this first attempt at improving Float Out leaves much to be
desired. In order to leave a marker for where to float a join point to, we make
a fake let binding for a special symbol, with the RHS being a made-up identifier
for the context. Besides being hacky, this means the RHS is a free occurrence of
an unknown identifier, which causes profuse warnings from CoreSubst.