CoreCxts: Use marginally less awful hack
We need some sort of marker in the code to represent the places beyond which a
join point may not float. Was using fake Let bindings for a special variable,
which is seven kinds of gross. Now instead the special variable goes on the RHS
of the fake binding so that the bound context identifier can have a properly
binding occurrence; this is only six kinds of gross.