Remove -fno-join-points
Now using master branch for comparison rather than varying the option. This flag
was a problematic way to do the comparison, partly because you really want to
compile the standard library with join points enabled, and if you do that,
-fno-join-points causes errors when things that *need* to be join points
(for instance, because it refers to some join point that came from an imported
unfolding) aren't made join points.