Find join points in occurrence analyser
This obsoletes the new CoreJoins module. Now the occurrence analyser changes
binders from values to join points where possible, if this is enabled.

The wrinkle here is that the code returned by the occurrence analyser is no
longer correct---the binding sites are changed to be join points, but none of
the occurrences are. In most ways, a disagreement between the binding site and
an occurrence is not a type error, but here it's a real problem. Thus CoreLint
now recognizes when it's looking at the output of the occurence analyser and
disables the check that a join id look like a join id at its occurrences.