Allow eta-contracted join points
Join points were required to be in eta-expanded form, largely for convenience
(though it very much *isn't* convenient in the simplifier, since it makes most
join points look non-trivial). The workaround is to have an accessor function
(the new CoreArity.splitJoinPoint) that eta-expands on demand for those
places that truly need it.