Handle type lets coming from desugarer
Now that simpleOptPgm sometimes creates type lets (because beta-redexes don't
work when join points are involved), a few more places have to deal with them.
In particular, if static pointers are on but optimization is off, Float Out runs
before the simplifier. Easy answer is to do a gentle simplifier run before Float
Out in this case.