Unbreak isRhsCtxt in OccurAnal
Since RHS analysis now always goes through occAnalLam, we need to make sure
occAnalLam preserves the original behavior when given an empty list of binders.
In particular, it should not change the environment in this case, since we need
to make sure it remains an RHS environment for the benefit of occAnalApp.