Keep track of join points in SimplIdSubst
See the Note. This actually reverts to an earlier situation, where CoreJoins
didn't change occurrence sites when making a binding into a join point. Now that
we instead rely on OccurAnal, which also doesn't change occurrences, we again
have to remember which binders refer to join points (whether or not they're in
the in-scope set).