Downgrade over-eager ASSERTs to WARNs
Was getting assertion failures on the following Core code, from CmmParse.hs:

    case (CmmType FloatCat W64)
         `cast` (sg_s10r2 ; UnsafeCo representational HappyAbsSyn Int
                 :: (CmmType :: *) ~R# (Int :: *))
    of wild_X5K [Dmd=<L,A>]
    { I# i_acTE [Dmd=<S,U>] ->
      ...
    }

This comes from Happy-generated code that uses manually tagged unions. It's from
a branch that's clearly bogus and would generate a runtime error, but won't
actually ever run (if Happy did its job right). With the ASSERTs in place, the
wip/join-points branch was failing to bootstrap because the above Core was
produced by the new case-of-join-point machinery.