Revert "Create join points when desugaring"
This reverts commit 4c6c257c6a68feb58f3fc4b8ad399e6df985d5e7.

Sadly, the change breaks many tests. (I seem to have run the testsuite from the
wrong checkout ...) The problem is nontrivial: The desugarer doesn't *just* use
"fail" variables as tail calls. It also passes them as arguments; in fact,
pattern synonyms desugar as essentially CPS functions, and they get passed
fails. So the fail ids can't be join points.

Alternatively, we could allow join points as parameters. This both is and isn't
difficult. In principle, we could allow labels to be passed as parameters *to
tail calls.* Then the called function could only use its join parameters by
tail-calling them, as usual.

Two big problems here:

1. If any of labels-as-parameters are around after Core Prep, we need codegen
and RTS support for this. Off the top of my head, one could use a pointer into
the stack to represent the label; the pointer would point at the return address
and so jumping would mean setting the stack pointer and then jumping where it
now points. But I'm sure the back-end guys would tell me where this is crazy.

2. It becomes necessary to know a function's exact arity (its "join arity,"
counting type arguments) so that we know what constitutes a tail call. This
turns arity from a squishy property to a hard invariant. At this point, we might
as well be implementing Strict Core or at least the "Types Are Calling
Conventions" paper, with its polyadic functions of fixed arity.