Count join points as cheap
We already had the -fcheap-join-points flag that counted each join binding as
free and each jump as 10 less than the function call would be. Simon suggested
counting the join binding *and its lambdas* as free and counting the jump as
free as well, so I implemented that. Now the only cost of a join point is the
cost of its right-hand side.

The NoFib results of this change were so good that I went ahead and removed the
option. Join points are now always cheap.