FloatOut: Separate unlifted value from rec group
If a recursive binding is floating to the same place as an unlifted let within
it, current behavior is to add the let to the rec group (because they get marked
with the same level). While this gets sorted out correctly by the occurrence
analyzer, it does break Core Lint to have an unlifted let marked as recursive.
This patch recognizes that the unlifted let must be ok-for-speculation and
therefore cannot actually be making a recursive call (despite appearing to
depend on the recursive group, if one goes by the level annotation). Hence we
separate these out.

This also solves a TODO by refactoring so that floatBind returns a list instead
of a single bind. We were already having to split up values and join points that
float to the same place; much cleaner to do this in floatBind.