SetLevels: Add context markers when decontifying
On the occasions when we're turning a join point back into a value, we end up
interpreting an incoming join point as a regular function (or thunk). This is
fine except that it's lacking the context marker that a value would have at the
top of the RHS, as well the one a function would have under the lambdas (if
any). Hence the join ceiling does not get set and remains at whatever value it
happend to have outside the RHS (which could be arbitrarily wrong), potentially
leading to incorrect floating decisions for join points.