Don't let worker/wrapper create bad join points
If a join point returns a function (hence has more lambdas than its join arity
suggests it should), we need to be very careful to leave it "eta-contracted",
taking only as many arguments as it needs to and passing them on (in W/W'd form)
to the worker. Otherwise the jump to the worker is buried under too many
lambdas (and has too many arguments).

This situation appears to be rare, so for the moment we just issue a warning (to
be sure it really is rare) and give up on W/W'ing the join point.